+++ app/src/main/java/com/ecxfoi/wbl/wienerbergerfrontend/ui/main/supporttickets/TicketListenerCallback.java
@@ -2,5 +2,5 @@ package com.ecxfoi.wbl.wienerbergerfrontend.ui.main.supporttickets;
 
 public interface TicketListenerCallback
 {
-    void onTicketSelected(int ticketId);
+    void onTicketSelected(Long ticketId);
 }
